### 2021-05-05 #3256 @whedon
[PRE REVIEW]: exoplanet: Gradient-based probabilistic inference forexoplanet data & other astronomical time series
**Submitting author:** @dfm (<a href="http://orcid.org/0000-0002-9328-5652">Daniel Foreman-Mackey</a>)
**Repository:** <a href="https://github.com/exoplanet-dev/exoplanet" target ="_blank">https://github.com/exoplanet-dev/exoplanet</a>
**Version:** v0.5.0
