+++ joss/paper.bib
@@ -350,17 +350,20 @@
 
 @article{guenther20,
   author        = {{G{\"u}nther}, Maximilian N. and {Daylan}, Tansu},
-  title         = {{Allesfitter: Flexible Star and Exoplanet Inference From Photometry and Radial Velocity}},
-  journal       = {arXiv e-prints},
-  keywords      = {Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
-  year          = 2020,
-  month         = mar,
-  eid           = {arXiv:2003.14371},
-  pages         = {arXiv:2003.14371},
+  title         = {{Allesfitter: Flexible Star and Exoplanet Inference from Photometry and Radial Velocity}},
+  journal       = {The Astrophysical Journal Supplement Series},
+  keywords      = {Exoplanets, Binary stars, Stellar flares, Bayesian statistics, Astronomy software, Starspots, Astronomy data modeling, 498, 154, 1603, 1900, 1855, 1572, 1859, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
+  year          = 2021,
+  month         = may,
+  volume        = {254},
+  number        = {1},
+  eid           = {13},
+  pages         = {13},
+  doi           = {10.3847/1538-4365/abe70e},
   archiveprefix = {arXiv},
   eprint        = {2003.14371},
   primaryclass  = {astro-ph.EP},
-  adsurl        = {https://ui.adsabs.harvard.edu/abs/2020arXiv200314371G},
+  adsurl        = {https://ui.adsabs.harvard.edu/abs/2021ApJS..254...13G},
   adsnote       = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
@@ -608,17 +611,20 @@
 
 @article{czekala21,
   author        = {{Czekala}, Ian and {Ribas}, {\'A}lvaro and {Cuello}, Nicol{\'a}s and {Chiang}, Eugene and {Mac{\'i}as}, Enrique and {Duch{\^e}ne}, Gaspard and {Andrews}, Sean M. and {Espaillat}, Catherine C.},
-  title         = {{A coplanar circumbinary protoplanetary disk in the TWA 3 triple M dwarf system}},
-  journal       = {arXiv e-prints},
-  keywords      = {Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
+  title         = {{A Coplanar Circumbinary Protoplanetary Disk in the TWA 3 Triple M Dwarf System}},
+  journal       = {The Astrophysical Journal},
+  keywords      = {Protoplanetary disks, Binary stars, Trinary stars, Stellar dynamics, Pre-main sequence stars, 1300, 154, 1714, 1596, 1290, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
   year          = 2021,
-  month         = feb,
-  eid           = {arXiv:2102.11875},
-  pages         = {arXiv:2102.11875},
+  month         = may,
+  volume        = {912},
+  number        = {1},
+  eid           = {6},
+  pages         = {6},
+  doi           = {10.3847/1538-4357/abebe3},
   archiveprefix = {arXiv},
   eprint        = {2102.11875},
   primaryclass  = {astro-ph.EP},
-  adsurl        = {https://ui.adsabs.harvard.edu/abs/2021arXiv210211875C},
+  adsurl        = {https://ui.adsabs.harvard.edu/abs/2021ApJ...912....6C},
   adsnote       = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
