### 2021-05-07 5885a8c
adding kitp ack and jax cite
