+++ joss/paper.bib
@@ -679,3 +679,11 @@
   adsurl   = {https://ui.adsabs.harvard.edu/abs/2017MNRAS.467.1702R},
   adsnote  = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@software{jax,
+  author  = {James Bradbury and Roy Frostig and Peter Hawkins and Matthew James Johnson and Chris Leary and Dougal Maclaurin and George Necula and Adam Paszke and Jake Vander{P}las and Skye Wanderman-{M}ilne and Qiao Zhang},
+  title   = {{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
+  url     = {http://github.com/google/jax},
+  version = {0.2.5},
+  year    = {2018}
+}
