### 2021-05-14 #3256 @arfon
:wave: @jbrande @nespinoza - would either of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission under review here is [_exoplanet: Gradient-based probabilistic inference forexoplanet data & other astronomical time series_](https://github.com/exoplanet-dev/exoplanet).
