### 2021-05-14 #3256 @arfon
@grburgess, @benjaminpope - many thanks for agreeing to review this submission for JOSS. See you over in #3285 where the actual review will take place.
