### 2021-05-14 #3285 @grburgess
@arfon I have had the same issue on other reviews... I am not able to check the boxes. It seems independent of the browser. Is this a common issue?
