### 2021-05-17 #3285 @benjaminpope
**Intro Comments:**
- Thanks for producing a great library, with many citations and applications already. This is already a very useful community resource. The code is well documented, with a host of great examples. I have read the paper and examples carefully, and tested several of these to the best of my poor laptop's ability. I recommend acceptance with minor revisions, providing a few comments below.

**Installation:**
- everything installs fine with pip and with git + `python setup.py`

**Unit tests:**
- All passed in my default Python environment except for `test_small_star` and `test_sky_coords`. This was because I hadn't yet installed batman - if these are going to be in unit tests perhaps they should also be in dependencies.

**Case Studies:**
- Ran all the case studies locally on my machine by copying and pasting code. A couple worked fine, the others seemed to work ok but had very long MCMC chains and I didn't hang around to check their output. Could be neat to have these as notebooks in a notebooks/ directory on GitHub, as you just know people (read: me) are going to hack these as recipes to run on their own data.
	- Just a note - check they all import matplotlib.pyplot as plt somewhere at the top? A couple of these don't and I needed to add this line. Perhaps the GitHub actions are importing this already?
	- It could also be nice to print run times on the case studies, so that prospective users know roughly how long each stage should take. Some of the MCMC chains and optimizations take a while to compute (of course), but as users are choosing between a few alternative packages, it will be helpful to know how this performs.
	- There are many warnings: perhaps in the interests of a reader hoping to clone code there could be a warnings filter placed at the top. For example, I get a lot of (and this may be to do with my Python environment):
		- `WARNING (theano.tensor.opt): Cannot construct a scalar test value from a test value with no size: InplaceDimShuffle{x,0}.0`
		- `WARNING (theano.tensor.opt): Cannot construct a scalar test value from a test value with no size: Elemwise{cos,no_inplace}.0`
- I'm actually a little uncertain about the boundary between Tutorial and Case Study. But I note that there is an excellent astrometry Tutorial (as long as the case studies and much longer than some of the tutorials), but no analogous Case Study doing astrometric fitting.
- _Gaussian process models for stellar variability_
	- Might like to add a sentence about why VI and MCMC retrieve such different posteriors?
- _Putting it all together:_
	- Shouldn't this come last? Currently second in the list.
- _RVs with multiple instruments_
	- Mine crashed on multiprocessing sampling. Perhaps default in any notebook version should be 1 core?

**General paper comments:**
- It is not unusual, though far from universal, for astronomy code in JOSS to be illustrated with a figure or two. (Around half of the recent papers I checked at random seem to have one). Perhaps the authors might like to include one in this paper, to illustrate the diverse capabilities of `exoplanet`. The multi-instrument light curve and RV fits from "Fitting light curves from multiple instruments" and "RVs with multiple instruments" respectively could go well as a multi-panel figure.  (This is by no means a requirement for the paper to be accepted, just an opportunity).
- The section on why autodiff is important in data analysis could perhaps do with some expansion. Not all astronomers are familiar with this, and perhaps another sentence or two selling and explaining the basic ideas of autodiff might help. (In my experience, many people do not realize it is different from finite differences). It might also do good to say explicitly whether it permits both forward and reverse mode autodiff, briefly highlight the advantages of Hamiltonian Monte Carlo, highlight the possibility of interoperability with neural network models, and so forth. I would even suggest putting a sentence in the Summary about the advantages of autodiff to hammer the point home. This is also not a required edit, but a suggestion to the authors on how to distinguish this work clearly.
- Citation formatting: fix brackets in "(based on the method used by the Stan project Carpenter et al.,2017.)" Similar problem with Theano, Theano Development Team, comma should be semicolon. Suggest semicolon before Carpenter would do?Should have citation to Jax.
- Citations - I know everyone has their homebrew exoplanet tool kit (which is why this project is so valuable) and that the Similar Tools section is is "far from a comprehensive list", but the paper could probably afford to cite more widely in this section. Perhaps a few other tools deserve a mention: for example Parviainen's PyExoTk or the Parviainen & Aigrain `ldtk` (I list these because they are the ones I have used myself; wide citation is to be encouraged, and perhaps the authors can think of others). Given the detached EB model, perhaps a citation to Irwin's `eb` or the `PHOEBE` or `JKTEBOP` projects?
- "finite-difference gradients which can be subject to significant numerical errors and require 2N computations of a models with N free parameters" - might be good to cite and quantify this statement, while contrasting the scaling behaviour with dimensionality of proper autodiff.
