### 2021-05-17 56423a8
make batman an optional dependency for tests
