### 2021-05-17 exoplanet-dev/exoplanet#187 @dfm
Add a documentation page giving an intro to autodiff
In [his JOSS review](https://github.com/openjournals/joss-reviews/issues/3285#issuecomment-842003491) @benjaminpope makes the excellent point that we have an opportunity to give more of an intro to autodiff and be a bit more explicit about why it is useful:

> The section on why autodiff is important in data analysis could perhaps do with some expansion. Not all astronomers are familiar with this, and perhaps another sentence or two selling and explaining the basic ideas of autodiff might help. (In my experience, many people do not realize it is different from finite differences). It might also do good to say explicitly whether it permits both forward and reverse mode autodiff, briefly highlight the advantages of Hamiltonian Monte Carlo, highlight the possibility of interoperability with neural network models, and so forth. I would even suggest putting a sentence in the Summary about the advantages of autodiff to hammer the point home. This is also not a required edit, but a suggestion to the authors on how to distinguish this work clearly.

I propose that this goes as a new page in the docs instead of in the paper because I believe that that would have higher impact.
