### 2021-05-17 exoplanet-dev/exoplanet#187 @benjaminpope
FYI I think your HMC lecture was very good, and I liked Gus Williams' intro to Jax: https://anguswilliams91.github.io/statistics/computing/jax/
