### 2021-05-17 exoplanet-dev/exoplanet#187 @dfm
Awesome - thanks for that reference, @benjaminpope!
