### 2021-05-17 e3227be
add a few more tool references
