+++ joss/paper.bib
@@ -687,3 +687,58 @@
   version = {0.2.5},
   year    = {2018}
 }
+
+@article{parviainen15b,
+  author        = {{Parviainen}, H. and {Aigrain}, S.},
+  title         = {{LDTK: Limb Darkening Toolkit}},
+  journal       = {Monthly Notices of the Royal Astronomical Society},
+  keywords      = {gravitational lensing: micro, methods: numerical, techniques: interferometric, planets and satellites: general, binaries: eclipsing, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
+  year          = 2015,
+  month         = nov,
+  volume        = {453},
+  number        = {4},
+  pages         = {3821-3826},
+  doi           = {10.1093/mnras/stv1857},
+  archiveprefix = {arXiv},
+  eprint        = {1508.02634},
+  primaryclass  = {astro-ph.EP},
+  adsurl        = {https://ui.adsabs.harvard.edu/abs/2015MNRAS.453.3821P},
+  adsnote       = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@article{irwin11,
+  author        = {{Irwin}, Jonathan M. and {Quinn}, Samuel N. and {Berta}, Zachory K. and {Latham}, David W. and {Torres}, Guillermo and {Burke}, Christopher J. and {Charbonneau}, David and {Dittmann}, Jason and {Esquerdo}, Gilbert A. and {Stefanik}, Robert P. and {Oksanen}, Arto and {Buchhave}, Lars A. and {Nutzman}, Philip and {Berlind}, Perry and {Calkins}, Michael L. and {Falco}, Emilio E.},
+  title         = {{LSPM J1112+7626: Detection of a 41 Day M-dwarf Eclipsing Binary from the MEarth Transit Survey}},
+  journal       = {The Astrophysical Journal},
+  keywords      = {binaries: eclipsing, stars: low-mass, Astrophysics - Solar and Stellar Astrophysics},
+  year          = 2011,
+  month         = dec,
+  volume        = {742},
+  number        = {2},
+  eid           = {123},
+  pages         = {123},
+  doi           = {10.1088/0004-637X/742/2/123},
+  archiveprefix = {arXiv},
+  eprint        = {1109.2055},
+  primaryclass  = {astro-ph.SR},
+  adsurl        = {https://ui.adsabs.harvard.edu/abs/2011ApJ...742..123I},
+  adsnote       = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@article{southworth04,
+  author        = {{Southworth}, J. and {Maxted}, P.~F.~L. and {Smalley}, B.},
+  title         = {{Eclipsing binaries in open clusters - II. V453 Cyg in NGC 6871}},
+  journal       = {Monthly Notices of the Royal Astronomical Society},
+  keywords      = {binaries: eclipsing, binaries: spectroscopic, stars: early-type, stars: fundamental parameters, stars: individual: V453 Cyg, open clusters and associations: individual: NGC 6871, Astrophysics},
+  year          = 2004,
+  month         = jul,
+  volume        = {351},
+  number        = {4},
+  pages         = {1277-1289},
+  doi           = {10.1111/j.1365-2966.2004.07871.x},
+  archiveprefix = {arXiv},
+  eprint        = {astro-ph/0403572},
+  primaryclass  = {astro-ph},
+  adsurl        = {https://ui.adsabs.harvard.edu/abs/2004MNRAS.351.1277S},
+  adsnote       = {Provided by the SAO/NASA Astrophysics Data System}
+}
