+++ joss/paper.bib
@@ -742,3 +742,22 @@
   adsurl        = {https://ui.adsabs.harvard.edu/abs/2004MNRAS.351.1277S},
   adsnote       = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@article{conroy20,
+  author        = {{Conroy}, Kyle E. and {Kochoska}, Angela and {Hey}, Daniel and {Pablo}, Herbert and {Hambleton}, Kelly M. and {Jones}, David and {Giammarco}, Joseph and {Abdul-Masih}, Michael and {Pr{\v{s}}a}, Andrej},
+  title         = {{Physics of Eclipsing Binaries. V. General Framework for Solving the Inverse Problem}},
+  journal       = {The Astrophysical Journal Supplement Series},
+  keywords      = {Eclipsing binary stars, 444, Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
+  year          = 2020,
+  month         = oct,
+  volume        = {250},
+  number        = {2},
+  eid           = {34},
+  pages         = {34},
+  doi           = {10.3847/1538-4365/abb4e2},
+  archiveprefix = {arXiv},
+  eprint        = {2006.16951},
+  primaryclass  = {astro-ph.SR},
+  adsurl        = {https://ui.adsabs.harvard.edu/abs/2020ApJS..250...34C},
+  adsnote       = {Provided by the SAO/NASA Astrophysics Data System}
+}
