### 2021-05-17 #3285 @dfm
Thanks @benjaminpope for your detailed and constructive review - all of these suggestions are very helpful!! As you've seen, I've created a series of issue threads to track these changes and I'll report back here once I've gone through them.
