### 2021-05-17 exoplanet-dev/exoplanet#188 @dfm
Move astrometry tutorial to case-studies and add a shorter astrometry tutorial
In https://github.com/openjournals/joss-reviews/issues/3285#issuecomment-842003491 @benjaminpope notes:

> I'm actually a little uncertain about the boundary between Tutorial and Case Study. But I note that there is an excellent astrometry Tutorial (as long as the case studies and much longer than some of the tutorials), but no analogous Case Study doing astrometric fitting.

This separation is really the result of the development history and I agree that it could be useful to clarify this difference. Zeroth order we should move the current astrometry tutorial (which mostly focuses on detailed data modeling points) to the case studies and add a simpler tutorial to these docs.

One thing that might be worth doing would be to just make *all* the tutorials only use simulated data and all the case studies use only real data (this would require moving the TTV case study to the tutorials which isn't a bad idea).
