### 2021-05-20 #3285 @grburgess
## general

This is a very well written package from a software point of view. It follows all the modern standards of CI and maintainable python code. The documentation appears complete and testing is well done.

I will second comments of @benjaminpope in order to not duplicate with an emphasis on expanding the autodiff explanation. Even more, why do you need HMC to do this? I wholeheartedly agree, but its a teaching opportunity to the field that these types of integrators are more robust for heavy-tailed and correlated posteriors and basically the only way to deal with high-dimensions. This seems to be a core benefit of this tool to the field, and could use a little more promotion.

I recommend publication as well, after addressing the issues I have linked above.

**pet peeve:** In the documentation there are several references to "finding the best parameters." I know what is heuristically meant, but perhaps one could use "estimating/computing the posterior" or "conditioning the model on data"? There is a perception in the field that Bayesian inference is a way to "find the parameter errors" more robustly than MLE methods, and careful use of language can help to overcome that.. but this is a style preference only :)

### Questions

* I noticed that one of the models was an interpolated light curve. Is this actually used in fitting? If it is used as a template where the shape does not change except for the scaling, this should be fine. But HMC typically goes bonkers with interpolated models that change their shape. This is more of curiosity of mine... but I also may have not understood what that model is used for in the library.

* As a non-expert, I see this as a framework that includes some models and can be built upon. Is this a correct interpretation? Have you experimented with different parameterizations of the models and/or have suggestions for users. HMC is great, but it can still be hampered by poor parameterization. I do note that scale parameters are "logged" as they should be (in general).
