### 2021-05-20 #3285 @dfm
@grburgess: Thanks for this review and the opened issues! Can I ask you to expand a bit on what you mean by:

> But HMC typically goes bonkers with interpolated models that **change their shape**.

And especially by "change their shape". I haven't found major issues with using interpolated models and I don't see any fundamental issue with using interpolation, so I'd love to hear more about your experiences!
