### 2021-05-20 #3285 @grburgess
@dfm In the past in Stan, when trying to use high-D interpolation, I've found that HMC gets really confused around the nodes. But, 1D interpolation, where the coefficients are precomputed and used like a template works fine. Example, you interpolate a light curve outside of the program from some complex model, and then are simply fitting for an amplitude that scales the interpolation. But I've had issues in the past were you also try to scale the time axis (stretching). I'm sure it has something to do with HMC diverging... but I never really investigated further.

 I was not sure what the intended use was here. If it is working that is great.
