### 2021-05-20 exoplanet-dev/exoplanet#196 @dfm
Audit docs for use of "best fit parameters"
From @grburgess in https://github.com/openjournals/joss-reviews/issues/3285#issuecomment-844827340

> In the documentation there are several references to "finding the best parameters." I know what is heuristically meant, but perhaps one could use "estimating/computing the posterior" or "conditioning the model on data"? There is a perception in the field that Bayesian inference is a way to "find the parameter errors" more robustly than MLE methods, and careful use of language can help to overcome that.. but this is a style preference only :)
