+++ setup.py
@@ -37,6 +37,7 @@ EXTRA_REQUIRE = {
         "pytest-cov>=2.6.1",
         "pytest-env",
         "batman-package",
+        "starry",
     ],
     "docs": [
         "sphinx>=1.7.5",
