+++ .github/workflows/tests.yml
@@ -4,11 +4,9 @@ on:
   push:
     branches:
       - main
+    tags:
+      - "*"
   pull_request:
-    branches:
-      - main
-  release:
-    types: [published]
 
 jobs:
   tests:
@@ -17,44 +15,31 @@ jobs:
       fail-fast: false
       matrix:
         python-version: ["3.9"]
-        pymc-arviz:
-          - '"pymc3~=3.9.0" "arviz<0.11"'
-          - '"pymc3~=3.10.0" "arviz<0.11"'
-          - '"pymc3~=3.11.0"'
-          - '"https://github.com/pymc-devs/pymc3/archive/master.zip"'
+        pymc-version: ["3.9", "3.10", "3.11", "dev"]
+      include:
+        - python-version: "3.8"
+          pymc-version: "3.11"
 
     steps:
-      - uses: actions/checkout@v2
+      - name: Checkout
+        uses: actions/checkout@v2
         with:
           fetch-depth: 0
-      - uses: actions/setup-python@v2
-        name: Install Python
+
+      - name: Setup Python
+        uses: actions/setup-python@v2
         with:
           python-version: ${{ matrix.python-version }}
 
       - name: Install dependencies
         run: |
           python -m pip install -U pip
-          python -m pip install ${{ matrix.pymc-arviz }}
-          python -m pip install -e .[test]
-          python -m pip install --no-deps starry
-
-      - name: Get compiledir
-        id: compiledir
-        shell: bash -l {0}
-        run: |
-          python -c "from aesara_theano_fallback import aesara; print('::set-output name=compiledir::' + aesara.config.compiledir)"
-
-      - name: Cache compiledir
-        uses: actions/cache@v2
-        with:
-          path: ${{ steps.compiledir.outputs.compiledir }}
-          key: compiledir-${{ steps.compiledir.outputs.compiledir }}-${{ matrix.pymc-arviz }}-${{ github.sha }}
-          restore-keys: |
-            compiledir-${{ steps.compiledir.outputs.compiledir }}-${{ matrix.pymc-arviz }}-
+          python -m pip install tox tox-gh-actions
 
       - name: Run tests
-        run: python -m pytest --cov=exoplanet -v tests
+        run: python -m tox
+        env:
+          PYMC_VERSION: ${{ matrix.pymc-version }}
 
       - name: Coveralls
         uses: AndreMiras/coveralls-python-action@v20201129
@@ -62,8 +47,27 @@ jobs:
           parallel: true
           flag-name: Unit Tests
 
+  lint:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+        with:
+          fetch-depth: 0
+
+      - name: Setup Python
+        uses: actions/setup-python@v2
+        with:
+          python-version: "3.9"
+
+      - name: Install dependencies
+        run: |
+          python -m pip install -U pip
+          python -m pip install tox
+
+      - name: Lint the code
+        run: python -m tox -e lint
+
   coverage:
-    name: "Merge and upload coverage"
     needs: tests
     runs-on: ubuntu-latest
     steps:
@@ -92,7 +96,7 @@ jobs:
           path: dist/*
 
   upload_pypi:
-    needs: [tests, build]
+    needs: [tests, lint, build]
     runs-on: ubuntu-latest
     if: github.event_name == 'release' && github.event.action == 'published'
     steps:
