+++ pytest.ini
@@ -1,4 +1,5 @@
 [pytest]
+testpaths = "tests"
 addopts =
     -v
 filterwarnings =
