+++ tox.ini
@@ -25,6 +25,7 @@ commands =
     python -m pytest -v {posargs}
 
 [testenv:lint]
+skip_install = true
 deps = pre-commit
 commands =
     pre-commit run --all-files
