+++ .github/workflows/tests.yml
@@ -16,9 +16,9 @@ jobs:
       matrix:
         python-version: ["3.9"]
         pymc-version: ["3.9", "3.10", "3.11", "dev"]
-      include:
-        - python-version: "3.8"
-          pymc-version: "3.11"
+        include:
+          - python-version: "3.8"
+            pymc-version: "3.11"
 
     steps:
       - name: Checkout
