### 2021-05-21 a5ef6b8
getting arviz versions working
