+++ tox.ini
@@ -15,8 +15,9 @@ PYMC_VERSION =
 
 [testenv]
 deps =
-    pymc39: pymc3~=3.9.0 arviz<0.11
-    pymc310: pymc3~=3.10.0 arviz<0.11
+    pymc{39,310}: arviz<0.11
+    pymc39: pymc3~=3.9.0
+    pymc310: pymc3~=3.10.0
     pymc311: pymc3~=3.11.0
     pymcdev: https://github.com/pymc-devs/pymc3/archive/v3.zip
 extras = test
