+++ environment.yml
@@ -1,12 +1,0 @@
-name: exoplanet
-channels:
-  - defaults
-dependencies:
-  - astropy
-  - pandas
-  - matplotlib
-  - pip
-  - python=3.7
-  - numpy
-  - pip:
-      - -e .[dev]
