+++ .github/workflows/tests.yml
@@ -34,18 +34,34 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install -U pip
-          python -m pip install tox tox-gh-actions
+          python -m pip install -U coveralls coverage[toml] tox tox-gh-actions
 
       - name: Run tests
         run: python -m tox
         env:
           PYMC_VERSION: ${{ matrix.pymc-version }}
 
-      - name: Coveralls
-        uses: AndreMiras/coveralls-python-action@v20201129
+      - name: Combine and upload coverage
+        run: |
+          python -m coverage combine
+          python -m coverage xml
+          python -m coveralls
+        env:
+          COVERALLS_PARALLEL: true
+
+  coverage:
+    needs: tests
+    runs-on: ubuntu-latest
+    steps:
+      - name: Setup Python
+        uses: actions/setup-python@v2
         with:
-          parallel: true
-          flag-name: Unit Tests
+          python-version: "3.9"
+      - name: Finish coverage collection
+        run: |
+          python -m pip install -U pip
+          python -m pip install -U coveralls
+          python -m coveralls --finish
 
   lint:
     runs-on: ubuntu-latest
@@ -66,15 +82,6 @@ jobs:
 
       - name: Lint the code
         run: python -m tox -e lint
-
-  coverage:
-    needs: tests
-    runs-on: ubuntu-latest
-    steps:
-      - name: Coveralls Finished
-        uses: AndreMiras/coveralls-python-action@v20201129
-        with:
-          parallel-finished: true
 
   build:
     runs-on: ubuntu-latest
