+++ tox.ini
@@ -14,6 +14,7 @@ PYMC_VERSION =
     dev: pymcdev
 
 [testenv]
+passenv = GITHUB_*
 deps =
     pymc{39,310}: arviz<0.11
     pymc39: pymc3~=3.9.0
