+++ setup.py
@@ -35,6 +35,7 @@ EXTRA_REQUIRE = {
         "scipy",
         "pytest",
         "pytest-cov>=2.6.1",
+        "coverage[toml]",
         "pytest-env",
         "batman-package",
     ],
