+++ pytest.ini
@@ -2,6 +2,7 @@
 testpaths = "tests"
 addopts =
     -v
+    --cov=exoplanet
 filterwarnings =
     ignore::DeprecationWarning
     ignore::FutureWarning
