+++ .github/workflows/tests.yml
@@ -44,6 +44,7 @@ jobs:
       - name: Combine and upload coverage
         run: python -m coveralls
         env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           COVERALLS_PARALLEL: true
 
   coverage:
