+++ pyproject.toml
@@ -15,10 +15,8 @@ use_parentheses = true
 known_third_party = ["astropy", "corner", "jupytext", "matplotlib", "nbconvert", "nbformat", "numpy", "packaging", "pandas", "pkg_resources", "pymc3", "pytest", "requests", "scipy", "setuptools", "sphinx_typlog_theme", "theano"]
 
 [tool.coverage.run]
-relative_files = true
 parallel = true
 branch = true
-source = ["exoplanet"]
 omit = [
   "*_test.py",
   "*__init__*",
@@ -38,7 +36,6 @@ source = ["src", ".tox/*/site-packages"]
 
 [tool.coverage.report]
 show_missing = true
-skip_covered = true
 exclude_lines = [
   "raise NotImplementedError",
   "except ImportError",
