+++ pyproject.toml
@@ -32,7 +32,7 @@ omit = [
 ]
 
 [tool.coverage.paths]
-source = ["src", ".tox/*/site-packages"]
+source = ["src", "*/site-packages"]
 
 [tool.coverage.report]
 show_missing = true
