+++ setup.py
@@ -34,7 +34,6 @@ EXTRA_REQUIRE = {
     "test": [
         "scipy",
         "pytest",
-        "pytest-cov>=2.6.1",
         "coverage[toml]",
         "pytest-env",
         "batman-package",
