+++ tox.ini
@@ -24,7 +24,7 @@ deps =
 extras = test
 commands =
     pip freeze
-    python -m pytest -v {posargs}
+    python -m coverage run -m pytest -v {posargs}
 
 [testenv:lint]
 skip_install = true
