+++ pytest.ini
@@ -2,7 +2,6 @@
 testpaths = "tests"
 addopts =
     -v
-    --cov=exoplanet
 filterwarnings =
     ignore::DeprecationWarning
     ignore::FutureWarning
