+++ .github/workflows/tests.yml
@@ -6,6 +6,9 @@ on:
       - main
     tags:
       - "*"
+    paths-ignore:
+      - "joss/**"
+      - "docs/**"
   pull_request:
 
 jobs:
