+++ .github/workflows/tutorials.yml
@@ -3,12 +3,11 @@ on:
   push:
     branches:
       - main
+    tags:
+      - "*"
+    paths-ignore:
+      - "joss/**"
   pull_request:
-    branches:
-      - main
-  release:
-    types:
-      - published
 
 jobs:
   notebooks:
