### 2021-05-22 f6e1638
Switching to tox for tests (#197)
* starting to switch to tox for tests

* tox tweaks

* Update tests.yml

* getting arviz versions working

* trying to get coverage working

* coverage settings

* coveralls

* be faster for now

* don't run combine

* token

* flags

* service

* conf

* dunno

* um?

* run all the tests again

* coverage

* skip version file

* ignore

* all tests

* omit properly

* path filters

* path filters
