+++ tox.ini
@@ -1,0 +1,33 @@
+[tox]
+envlist = py{38,39}-pymc{39,310,311,dev},lint
+
+[gh-actions]
+python =
+    3.8: py38
+    3.9: py39
+
+[gh-actions:env]
+PYMC_VERSION =
+    3.9: pymc39
+    3.10: pymc310
+    3.11: pymc311
+    dev: pymcdev
+
+[testenv]
+passenv = GITHUB_*
+deps =
+    pymc{39,310}: arviz<0.11
+    pymc39: pymc3~=3.9.0
+    pymc310: pymc3~=3.10.0
+    pymc311: pymc3~=3.11.0
+    pymcdev: https://github.com/pymc-devs/pymc3/archive/v3.zip
+extras = test
+commands =
+    pip freeze
+    python -m coverage run -m pytest -v {posargs}
+
+[testenv:lint]
+skip_install = true
+deps = pre-commit
+commands =
+    pre-commit run --all-files
