+++ tests/distributions_test.py
@@ -1,12 +1,17 @@
 # -*- coding: utf-8 -*-
 
 import logging
+from collections import namedtuple
 
+import aesara_theano_fallback.tensor as tt
 import numpy as np
 import pymc3 as pm
 import pytest
+from pymc3.tests.test_distributions import R, Unit, Vector
+from pymc3.tests.test_transforms import check_transform, get_values
 from scipy.stats import beta, halfnorm, kstest, rayleigh
 
+from exoplanet.distributions import transforms as tr
 from exoplanet.distributions.eccentricity import kipping13, vaneylen19
 from exoplanet.distributions.physical import ImpactParameter, QuadLimbDark
 
@@ -222,3 +227,28 @@ class TestPhysical(_Base):
             assert s < 0.05
 
         assert np.all(trace["b"] <= 1 + trace["ror"])
+
+
+def test_quad_limb_dark_transform():
+    values = get_values(
+        tr.quad_limb_dark,
+        Vector(R, 2),
+        constructor=tt.vector,
+        test=np.array([0.0, 0.0]),
+    )
+    domain = namedtuple("Domain", ["vals"])(values)
+    check_transform(
+        tr.quad_limb_dark,
+        domain,
+        constructor=tt.vector,
+        test=np.array([0.0, 0.0]),
+    )
+
+
+def test_impact_parameter_transform():
+    ror = np.float64(0.03)
+    check_transform(
+        tr.impact_parameter(ror),
+        Unit * (1 + ror),
+        test=0.5,
+    )
