+++ src/exoplanet/orbits/dur_to_ecc.py
@@ -11,9 +11,9 @@ import pymc3 as pm
 from .keplerian import KeplerianOrbit, _get_consistent_inputs
 
 
-def duration_to_eccentricity(  # pragma: no cover
+def duration_to_eccentricity(
     func, duration, ror, **kwargs
-):
+):  # pragma: no cover
     num_planets = kwargs.pop("num_planets", 1)
     orbit_type = kwargs.pop("orbit_type", KeplerianOrbit)
     name = kwargs.get("name", "dur_ecc")
