### 2021-05-23 1c78502
Adding a bit more coverage (#199)
* adding a bit more coverage

* no cover syntax
