### 2021-05-25 #3285 @grburgess
@arfon I am still unable to interact with the checkboxes.
