+++ docs/tutorials/notebook_setup.py
@@ -1,50 +1,0 @@
-# ---
-# jupyter:
-#   jupytext:
-#     text_representation:
-#       extension: .py
-#       format_name: light
-#       format_version: '1.5'
-#       jupytext_version: 1.7.1
-#   kernelspec:
-#     display_name: Python 3
-#     language: python
-#     name: python3
-# ---
-
-# +
-"""isort:skip_file"""
-
-get_ipython().magic('config InlineBackend.figure_format = "retina"')
-
-import os
-import logging
-import warnings
-
-import matplotlib.pyplot as plt
-
-# Don't use the schmantzy progress bar
-os.environ["EXOPLANET_NO_AUTO_PBAR"] = "true"
-
-# Remove when Theano is updated
-warnings.filterwarnings("ignore", category=DeprecationWarning)
-warnings.filterwarnings("ignore", category=FutureWarning)
-
-# Remove when arviz is updated
-warnings.filterwarnings("ignore", category=UserWarning)
-
-
-logger = logging.getLogger("theano.gof.compilelock")
-logger.setLevel(logging.ERROR)
-logger = logging.getLogger("exoplanet")
-logger.setLevel(logging.DEBUG)
-
-
-plt.style.use("default")
-plt.rcParams["savefig.dpi"] = 100
-plt.rcParams["figure.dpi"] = 100
-plt.rcParams["font.size"] = 16
-plt.rcParams["font.family"] = "sans-serif"
-plt.rcParams["font.sans-serif"] = ["Liberation Sans"]
-plt.rcParams["font.cursive"] = ["Liberation Sans"]
-plt.rcParams["mathtext.fontset"] = "custom"

[…]
