+++ docs/tutorials/transit.md
@@ -1,273 +1,0 @@
----
-jupytext:
-  text_representation:
-    extension: .md
-    format_name: myst
-    format_version: 0.13
-    jupytext_version: 1.11.2
-kernelspec:
-  display_name: Python 3
-  language: python
-  name: python3
----
-
-```{code-cell}
-:nbsphinx: hidden
-
-%matplotlib inline
-```
-
-```{code-cell}
-:nbsphinx: hidden
-
-%run notebook_setup
-```
-
-# Transit fitting
-
-+++
-
-*exoplanet* includes methods for computing the light curves transiting planets.
-In its simplest form this can be used to evaluate a light curve like you would do with [batman](https://astro.uchicago.edu/~kreidberg/batman/), for example:
-
-```{code-cell}
-import numpy as np
-import matplotlib.pyplot as plt
-
-import exoplanet as xo
-
-# The light curve calculation requires an orbit
-orbit = xo.orbits.KeplerianOrbit(period=3.456)
-
-# Compute a limb-darkened light curve using starry
-t = np.linspace(-0.1, 0.1, 1000)
-u = [0.3, 0.2]
-light_curve = (
-    xo.LimbDarkLightCurve(*u)
-    .get_light_curve(orbit=orbit, r=0.1, t=t, texp=0.02)
-    .eval()
-)
-# Note: the `eval` is needed because this is using Theano in
-# the background
-
-plt.plot(t, light_curve, color="C0", lw=2)
-plt.ylabel("relative flux")
-plt.xlabel("time [days]")
-_ = plt.xlim(t.min(), t.max())
-```
-
-But the real power comes from the fact that this is defined as a [Theano operation](https://aesara.readthedocs.io/en/latest/extending/index.html) so it can be combined with PyMC3 to do transit inference using Hamiltonian Monte Carlo.
-
-## The transit model in PyMC3
-
-In this section, we will construct a simple transit fit model using *PyMC3* and then we will fit a two planet model to simulated data.
-To start, let's randomly sample some periods and phases and then define the time sampling:
-
-```{code-cell}
-np.random.seed(123)
-periods = np.random.uniform(5, 20, 2)
-t0s = periods * np.random.rand(2)
-t = np.arange(0, 80, 0.02)
-yerr = 5e-4
-```
-
-Then, define the parameters.
-In this simple model, we'll just fit for the limb darkening parameters of the star, and the period, phase, impact parameter, and radius ratio of the planets (note: this is already 10 parameters and running MCMC to convergence using [emcee](https://emcee.readthedocs.io) would probably take at least an hour).
-For the limb darkening, we'll use a quadratic law as parameterized by [Kipping (2013)](https://arxiv.org/abs/1308.0009).
-This reparameterizations is implemented in *exoplanet* as custom *PyMC3* distribution :class:`exoplanet.distributions.QuadLimbDark`.
-
-```{code-cell}
-import pymc3 as pm
-import pymc3_ext as pmx
-
-with pm.Model() as model:
-
-    # The baseline flux
-    mean = pm.Normal("mean", mu=0.0, sd=1.0)
-
-    # The time of a reference transit for each planet
-    t0 = pm.Normal("t0", mu=t0s, sd=1.0, shape=2)
-
-    # The log period; also tracking the period itself
-    logP = pm.Normal("logP", mu=np.log(periods), sd=0.1, shape=2)
-    period = pm.Deterministic("period", pm.math.exp(logP))
-
-    # The Kipping (2013) parameterization for quadratic limb darkening paramters
-    u = xo.distributions.QuadLimbDark("u", testval=np.array([0.3, 0.2]))
-
-    r = pm.Uniform(
[…]
