+++ docs/tutorials/why.md
@@ -1,0 +1,22 @@
+---
+jupytext:
+  text_representation:
+    extension: .md
+    format_name: myst
+    format_version: 0.13
+    jupytext_version: 1.11.2
+kernelspec:
+  display_name: Python 3
+  language: python
+  name: python3
+---
+
+# Why does this exist?
+
++++
+
+This section will try to motivate the point of exoplanet and gradient-based inference.
+
+```{code-cell}
+
+```
