+++ docs/tutorials/data-and-models.md
@@ -1,0 +1,22 @@
+---
+jupytext:
+  text_representation:
+    extension: .md
+    format_name: myst
+    format_version: 0.13
+    jupytext_version: 1.11.2
+kernelspec:
+  display_name: Python 3
+  language: python
+  name: python3
+---
+
+# Overview of the supported data & models
+
++++
+
+This will give intros to transit, rv, and astrometric planet models. Talk about combining datasets, shared parameters, joint fits, and GPs?
+
+```{code-cell}
+
+```
