+++ docs/tutorials/about.md
@@ -1,0 +1,24 @@
+---
+jupytext:
+  text_representation:
+    extension: .md
+    format_name: myst
+    format_version: 0.13
+    jupytext_version: 1.11.2
+kernelspec:
+  display_name: Python 3
+  language: python
+  name: python3
+---
+
+# About these tutorials
+
++++
+
+This section will describe these tutorials, how to use them, what they're about, the case studies, etc.
+
+Talk about runtime?
+
+```{code-cell}
+
+```
