+++ docs/tutorials/autodiff.md
@@ -1,0 +1,22 @@
+---
+jupytext:
+  text_representation:
+    extension: .md
+    format_name: myst
+    format_version: 0.13
+    jupytext_version: 1.11.2
+kernelspec:
+  display_name: Python 3
+  language: python
+  name: python3
+---
+
+# Autodiff & gradient-based inference
+
++++
+
+This section will try to motivate the point of exoplanet, gradient-based inference, and intro to autodiff.
+
+```{code-cell}
+
+```
