### 2021-05-31 #3285 @benjaminpope
Re Whedon post - just waiting for confirmation on all the pull requests that changes have been made. Everything looks good so far - the comment about paper content & adding a figure (which is not compulsory at all) hasn't been decided either way, so as far as I can tell I'm just waiting on that.
