+++ .pre-commit-config.yaml
@@ -22,7 +22,7 @@ repos:
     exclude: docs/tutorials
 
 - repo: https://github.com/psf/black
-  rev: "21.5b1"
+  rev: "21.5b2"
   hooks:
   - id: black
 
