+++ readthedocs.yml
@@ -4,7 +4,7 @@ submodules:
   include: all
 
 python:
-  version: 3.7
+  version: 3.8
   install:
     - method: pip
       path: .
@@ -12,4 +12,6 @@ python:
         - docs
 
 sphinx:
+  configuration: docs/conf.py
+  fail_on_warning: true
   builder: dirhtml
