+++ .github/workflows/tutorials.yml
@@ -1,62 +1,0 @@
-name: Tutorials
-on:
-  push:
-    branches:
-      - main
-    tags:
-      - "*"
-    paths-ignore:
-      - "joss/**"
-  pull_request:
-
-jobs:
-  notebooks:
-    name: "Build the notebooks for the docs"
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-        with:
-          submodules: true
-
-      - name: Set up Python
-        uses: actions/setup-python@v2
-        with:
-          python-version: 3.9
-
-      - name: Install dependencies
-        run: |
-          python -m pip install -U pip
-          python -m pip install "pymc3~=3.10"
-          python -m pip install ".[tutorials]"
-
-      - name: Get compiledir
-        id: compiledir
-        shell: bash -l {0}
-        run: |
-          python -c "from aesara_theano_fallback import aesara; print('::set-output name=compiledir::' + aesara.config.compiledir)"
-
-      - name: Cache compiledir
-        uses: actions/cache@v2
-        with:
-          path: ${{ steps.compiledir.outputs.compiledir }}
-          key: tutorials-${{ steps.compiledir.outputs.compiledir }}-${{ github.sha }}
-          restore-keys: |
-            tutorials-${{ steps.compiledir.outputs.compiledir }}-
-            tutorials-
-
-      - name: Execute the notebooks
-        run: |
-          jupytext --to ipynb --execute docs/tutorials/*.py
-
-      - uses: actions/upload-artifact@v2
-        with:
-          name: notebooks-for-${{ github.sha }}
-          path: docs/tutorials
-
-      - name: Trigger RTDs build
-        if: ${{ github.event_name != 'pull_request' }}
-        uses: dfm/rtds-action@v1.0.3
-        with:
-          webhook_url: ${{ secrets.RTDS_WEBHOOK_URL }}
-          webhook_token: ${{ secrets.RTDS_WEBHOOK_TOKEN }}
-          commit_ref: ${{ github.ref }}
