+++ docs/tutorials/about.md
@@ -30,4 +30,8 @@ print(f"exoplanet.__version__ = '{exoplanet.__version__}'")
 
 That cell also includes a call to the `exoplanet.utils.docs_setup` function that will squash some warnings (these are generally caused by Theano/Aesara; see {ref}`theano` for more info) and set up our `matplotlib` style.
 
-To run the on your own, you can click on the
+To exectute a tutorial on your own, you can click on the buttons at the top right or this page to launch the notebook using [Binder](https://mybinder.org) or download the `.ipynb` file directly.
+
+```{code-cell}
+
+```
