### 2021-06-01 c204043
don't fail on warning for now
