+++ readthedocs.yml
@@ -13,5 +13,5 @@ python:
 
 sphinx:
   configuration: docs/conf.py
-  fail_on_warning: true
+  # fail_on_warning: true
   builder: dirhtml
