+++ readthedocs.yml
@@ -14,4 +14,4 @@ python:
 sphinx:
   configuration: docs/conf.py
   # fail_on_warning: true
-  builder: dirhtml
+  builder: html
