### 2021-06-01 604e01c
adding rtds conda env
