+++ readthedocs.yml
@@ -3,13 +3,16 @@ version: 2
 submodules:
   include: all
 
-python:
-  version: 3.8
-  install:
-    - method: pip
-      path: .
-      extra_requirements:
-        - docs
+conda:
+  environment: docs/environment.yml
+
+# python:
+#   version: 3.8
+#   install:
+#     - method: pip
+#       path: .
+#       extra_requirements:
+#         - docs
 
 sphinx:
   configuration: docs/conf.py
