### 2021-06-01 fd6ef74
more specific conda reqs [ci skip]
