+++ readthedocs.yml
@@ -6,15 +6,7 @@ submodules:
 conda:
   environment: docs/environment.yml
 
-# python:
-#   version: 3.8
-#   install:
-#     - method: pip
-#       path: .
-#       extra_requirements:
-#         - docs
-
 sphinx:
   configuration: docs/conf.py
-  # fail_on_warning: true
+  fail_on_warning: true
   builder: html
