+++ docs/tutorials/autodiff.md
@@ -11,6 +11,8 @@ kernelspec:
   name: python3
 ---
 
+(autodiff)=
+
 # Automatic differentation & gradient-based inference
 
 ```{code-cell}
@@ -24,11 +26,35 @@ The major selling point of `exoplanet` compared to other similar libraries is th
 `PyMC3` offers gradient-based inference algorithms (more on these below) that can be more computationally efficient (per effective sample) than other tools commonly used for probabilistic inference in astrophysics.
 When I am writing this, gradient-based inference methodology is not widely used in astro because it can be difficult to compute the relevant gradients (derivatives of your log probability function with respect to the parameters).
 `PyMC3` (and many other frameworks like it) handle this issue using [automatic differentiation (AD)](https://en.wikipedia.org/wiki/Automatic_differentiation), a method (or collection of methods) for automatically propagating derivatives through your code.
+
 It's beyond the scope of this tutorial to go into too many details about AD and most users of `exoplanet` shouldn't need to interact with this too much, but this should at least give you a little taste of the kinds of things AD can do for you and demonstrate how this translates into efficient inference with probabilistic models.
+The main thing that I want to emphasize here is that AD is not the same as *symbolic differentiation* (it's not going to provide you with a mathematical expression for your gradients), but it's also not the same as numerical methods like [finite difference](https://en.wikipedia.org/wiki/Finite_difference).
+Using AD to evaluate the gradients of your model will generally be faster, more efficient, and more numerically stable than alternatives, but there are always exceptions to any rule.
+There are times when providing your AD framework with a custom implementation and/or differentation rule for a particular function is beneficial in terms of cost and stability.
+`exoplanet` is designed to provide these custom implementations only where it is useful (e.g. solving Kepler's equation or evaluating limb-darkened light curves) and then rely on the existing AD toolkit elsewhere.
 
 ## Automatic differentation in Theano/Aesara
 
-`PyMC3` is built on top of Theano/Aesara.
+Let's start by giving a quick taste for how AD works in the context of `PyMC3` and `exoplanet`.
+`PyMC3` is built on top of a library called `Aesara` (formerly `Theano`, see {ref}`theano`) that was designed for AD and fast linear algebra in the context of neural networks.
+The learning curve can be steep because your code ends up looking a little different from the usual way you would write Python code.
+(If you're looking for something more "Pythonic", [JAX](https://github.com/google/jax) might be a good alternative, but for our purposes we'll stick with `Aesara`.)
+In `Aesara`, you start be defining the *relationships* between variables and then ask to evaluate a variable using this "graph".
+This might seem a little counterintuitive and you hopefully won't have to worry about this too much, but it can be a handy workflow because `Aesara` will compile and optimize your model in the background (generally making it faster).
+This structure allows us to evaluate the derivative of some variable with respect to another.
+
+Here's a simple example where we define a relationship between two variables `x` and `y`:
+
+$$
+y = \exp\left[\sin\left(\frac{2\,\pi\,x}{3}\right)\right]
+$$
+
+and then calculate the derivative using AD.
+For comparison, the symbolic derivative is:
+
+$$
[…]
