+++ docs/tutorials/data-and-models.md
@@ -11,11 +11,226 @@ kernelspec:
   name: python3
 ---
 
-# Overview of the supported data & models
+(data-and-models)=
+
+# Data & models
+
+```{code-cell}
+import exoplanet
+
+exoplanet.utils.docs_setup()
+print(f"exoplanet.__version__ = '{exoplanet.__version__}'")
+```
+
+This tutorial provides an overview of the kinds of datasets that `exoplanet` has been designed to model.
+It's worth noting that this is certainly not a comprehensive list and `exoplanet` is designed to be *low-level* library that can be built on top of to support new use cases.
+
+:::{tip}
+For real world use cases and examples, check out the [Case Studies](https://gallery.exoplanet.codes) page.
+:::
 
 +++
 
-This will give intros to transit, rv, and astrometric planet models. Talk about combining datasets, shared parameters, joint fits, and GPs?
+## Overview
+
+One primary interface to `exoplanet` is via the {class}`exoplanet.orbits.KeplerianOrbit` object (and its generalizations like {class}`exoplanet.orbits.TTVOrbit`).
+The `KeplerianOrbit` object defines a set of independent (but possibly massive) bodies orbiting a central body on Keplerian trajectories.
+These orbits are parameterized by orbital elements, and one of the primary uses of `exoplanet` is the measurement of these orbital elements given observations.
+
+Given a `KeplerianOrbit`, users of `exoplanet` can evaluate things like the positions and velocities of all the bodies as a function of time.
+For example, here's how you could define an orbit with a single body and plot the orbits:
+
+```{code-cell}
+import numpy as np
+import matplotlib.pyplot as plt
+import exoplanet as xo
+
+# Define the orbit
+orbit = xo.orbits.KeplerianOrbit(
+    period=10.0,  # All times are in days
+    t0=0.5,
+    incl=0.5 * np.pi,  # All angles are in radians
+    ecc=0.3,
+    omega=-2.5,
+    Omega=1.2,
+    m_planet=0.05,  # All masses and distances are in Solar units
+)
+
+# Get the position and velocity as a function of time
+t = np.linspace(0, 20, 5000)
+x, y, z = orbit.get_relative_position(t)
+vx, vy, vz = orbit.get_star_velocity(t)
+
+# Plot the coordinates
+# Note the use of `.eval()` throughout since the coordinates are all
+# Aesara/Theano objects
+fig, axes = plt.subplots(2, 1, figsize=(8, 8), sharex=True)
+ax = axes[0]
+ax.plot(t, x.eval(), label="x")
+ax.plot(t, y.eval(), label="y")
+ax.plot(t, z.eval(), label="z")
+ax.set_ylabel("position of orbiting body [$R_*$]")
+ax.legend(fontsize=10, loc=1)
+
+ax = axes[1]
+ax.plot(t, vx.eval(), label="$v_x$")
+ax.plot(t, vy.eval(), label="$v_y$")
+ax.plot(t, vz.eval(), label="$v_z$")
+ax.set_xlim(t.min(), t.max())
+ax.set_xlabel("time [days]")
+ax.set_ylabel("velocity of central [$R_*$/day]")
+_ = ax.legend(fontsize=10, loc=1)
+```
+
+The key feture of `exoplanet` is that all of the parameters to a `KeplerianOrbit` can be `PyMC3` variables.
+This means that these elements are now something that you can *infer*.
+For example, if we want to fit for the orbital period, we can define a `PyMC3` model like the following:
+
+```{code-cell}
+import pymc3 as pm
+
+with pm.Model():
+    log_period = pm.Normal("log_period", mu=np.log(10), sigma=2.0)
+    orbit = xo.orbits.KeplerianOrbit(
+        period=pm.math.exp(log_period),  # ...
+    )
+
[…]
