+++ docs/tutorials/data-and-models.md
@@ -218,17 +218,184 @@ For example, if the orbit is defined using the period $P$, the semi-major axis $
 
 +++
 
+(data-and-models/rvs)=
+
 ## Radial velocities
 
+A {class}`exoplanet.orbits.KeplerianOrbit` can be used to compute the expected radial velocity time series for a given set of parameters.
+One typical parameterization for a radial velocity fit would look something like this:
+
+```{code-cell}
+import arviz as az
+import pymc3_ext as pmx
+import aesara_theano_fallback.tensor as tt
+
+# Create a dummy dataset
+random = np.random.default_rng(1234)
+t = np.sort(random.uniform(0, 50, 20))
+rv_err = 0.01 + np.zeros_like(t)
+rv_obs = 5.0 * np.sin(2 * np.pi * t / 10.0) + np.sqrt(
+    0.05 ** 2 + rv_err ** 2
+) * random.normal(size=len(t))
+
+with pm.Model():
+
+    # Period, semi-amplitude, and eccentricity
+    log_period = pm.Normal("log_period", mu=np.log(10.0), sigma=1.0)
+    period = pm.Deterministic("period", tt.exp(log_period))
+    log_semiamp = pm.Normal("log_semiamp", mu=np.log(5.0), sd=2.0)
+    semiamp = pm.Deterministic("semiamp", tt.exp(log_semiamp))
+    ecc = pm.Uniform("ecc", lower=0, upper=1)
+
+    # At low eccentricity, omega and the phase of periastron (phi) are
+    # correlated so it can be best to fit in (omega ± phi) / 2
+    plus = pmx.Angle("plus")
+    minus = pmx.Angle("minus")
+    phi = pm.Deterministic("phi", plus + minus)
+    omega = pm.Deterministic("omega", plus - minus)
+
+    # For non-zero eccentricity, it can sometimes be better to use
+    # sqrt(e)*sin(omega) and sqrt(e)*cos(omega) as your parameters:
+    #
+    #     ecs = pmx.UnitDisk("ecs", testval=0.01 * np.ones(2))
+    #     ecc = pm.Deterministic("ecc", tt.sum(ecs ** 2, axis=0))
+    #     omega = pm.Deterministic("omega", tt.arctan2(ecs[1], ecs[0]))
+    #     phi = pmx.Angle("phi")
+
+    # Jitter & the system mean velocity offset
+    log_jitter = pm.Normal("log_jitter", mu=np.log(0.05), sd=5.0)
+    zero_point = pm.Normal("zero_point", mu=0, sd=10.0)
+
+    # Then we define the orbit
+    tperi = pm.Deterministic("tperi", period * phi / (2 * np.pi))
+    orbit = xo.orbits.KeplerianOrbit(
+        period=period, t_periastron=tperi, ecc=ecc, omega=omega
+    )
+
+    # And define the RV model
+    rv_model = zero_point + orbit.get_radial_velocity(t, K=semiamp)
+
+    # Finally add in the observation model
+    err = tt.sqrt(rv_err ** 2 + tt.exp(2 * log_jitter))
+    pm.Normal("obs", mu=rv_model, sigma=rv_err, observed=rv_obs)
+
+    soln = pmx.optimize(vars=[plus, minus, ecc])
+    soln = pmx.optimize(soln)
+    trace = pmx.sample(
+        tune=1000,
+        draws=1000,
+        cores=2,
+        chains=2,
+        start=soln,
+        return_inferencedata=True,
+    )
+
+az.summary(trace)
+```
+
+Instead of using `semiamp` as a parameter, we could have passed `m_planet` and `incl` as parameters to `KeplerianOrbit` and fit the physical orbit.
[…]
