+++ readthedocs.yml
@@ -9,4 +9,4 @@ conda:
 sphinx:
   configuration: docs/conf.py
   fail_on_warning: true
-  builder: html
+  builder: dirhtml
