### 2021-06-08 cb0b1f9
words about reparam and combining
