### 2021-06-08 787ad30
Major refactor of tutorials (#200)
* switching to markdown

* adding some placeholders for new tutorials

* pre-commit setup

* pre-commit setup

* pre-commit setup

* pre-commit setup

* add a few more placeholders

* formatting

* running tutorials on RTDs

* testing docs generated with tox

* more words in about doc

* not much

* don't fail on warning for now

* try html build

* adding rtds conda env

* syntax [ci skip]

* more specific conda reqs [ci skip]

* a few words

* words in autodiff tutorial

* some words about orbital conventions

* rv and astrometry

* plots in data/models sec
[…]
