+++ docs/tutorials/intro-to-pymc3.py
@@ -1,404 +1,0 @@
-# -*- coding: utf-8 -*-
-# ---
-# jupyter:
-#   jupytext:
-#     text_representation:
-#       extension: .py
-#       format_name: light
-#       format_version: '1.5'
-#       jupytext_version: 1.7.1
-#   kernelspec:
-#     display_name: Python 3
-#     language: python
-#     name: python3
-# ---
-
-# + nbsphinx="hidden"
-# %matplotlib inline
-
-# + nbsphinx="hidden"
-# %run notebook_setup
-# -
-
-# # A quick intro to PyMC3 for exoplaneteers
-
-# [Hamiltonian Monte Carlo (HMC)](https://en.wikipedia.org/wiki/Hamiltonian_Monte_Carlo) methods haven't been widely used in astrophysics, but they are the standard methods for probabilistic inference using Markov chain Monte Carlo (MCMC) in many other fields.
-# *exoplanet* is designed to provide the building blocks for fitting many exoplanet datasets using this technology, and this tutorial presents some of the basic features of the [PyMC3](https://docs.pymc.io/) modeling language and inference engine.
-# The [documentation for PyMC3](https://docs.pymc.io/) includes many other tutorials that you should check out to get more familiar with the features that are available.
-#
-# In this tutorial, we will go through two simple examples of fitting some data using PyMC3.
-# The first is the classic fitting a line to data with unknown error bars, and the second is a more relevant example where we fit a radial velocity model to the public radial velocity observations of [51 Peg](https://en.wikipedia.org/wiki/51_Pegasi).
-# You can read more about fitting lines to data [in the bible of line fitting](https://arxiv.org/abs/1008.4686) and you can see another example of fitting the 51 Peg data using HMC (this time using [Stan](http://mc-stan.org)) [here](https://dfm.io/posts/stan-c++/).
-#
-# ## Hello world (AKA fitting a line to data)
-#
-# My standard intro to a new modeling language or inference framework is to fit a line to data.
-# So. Let's do that with PyMC3.
-#
-# To start, we'll generate some fake data using a linear model.
-# Feel free to change the random number seed to try out a different dataset.
-
-# +
-import numpy as np
-import matplotlib.pyplot as plt
-
-np.random.seed(42)
-
-true_m = 0.5
-true_b = -1.3
-true_logs = np.log(0.3)
-
-x = np.sort(np.random.uniform(0, 5, 50))
-y = true_b + true_m * x + np.exp(true_logs) * np.random.randn(len(x))
-
-plt.plot(x, y, ".k")
-plt.ylim(-2, 2)
-plt.xlabel("x")
-plt.ylabel("y")
-# -
-
-# To fit a model to these data, our model will have 3 parameters: the slope $m$, the intercept $b$, and the log of the uncertainty $\log(\sigma)$.
-# To start, let's choose broad uniform priors on these parameters:
-#
-# $$
-# \begin{eqnarray}
-# p(m) &=& \left\{\begin{array}{ll}
-# 1/10 & \mathrm{if}\,-5 < m < 5 \\
-# 0 & \mathrm{otherwise} \\
-# \end{array}\right. \\
-# p(b) &=& \left\{\begin{array}{ll}
-# 1/10 & \mathrm{if}\,-5 < b < 5 \\
-# 0 & \mathrm{otherwise} \\
-# \end{array}\right. \\
-# p(\log(\sigma)) &=& \left\{\begin{array}{ll}
-# 1/10 & \mathrm{if}\,-5 < \log(\sigma) < 5 \\
[…]
