+++ src/exoplanet/utils.py
@@ -1,6 +1,12 @@
 # -*- coding: utf-8 -*-
 
-__all__ = ["logger", "as_tensor_variable", "deprecation_warning", "deprecated"]
+__all__ = [
+    "logger",
+    "as_tensor_variable",
+    "deprecation_warning",
+    "deprecated",
+    "docs_setup",
+]
 
 import logging
 import warnings
@@ -36,3 +42,34 @@ def deprecated(alternate=None):  # pragma: no cover
         return f
 
     return wrapper
+
+
+def docs_setup():
+    """Set some environment variables and ignore some warnings for the docs"""
+    import logging
+    import warnings
+
+    import matplotlib.pyplot as plt
+
+    # Remove when Theano is updated
+    warnings.filterwarnings("ignore", category=DeprecationWarning)
+    warnings.filterwarnings("ignore", category=FutureWarning)
+
+    # Remove when arviz is updated
+    warnings.filterwarnings("ignore", category=UserWarning)
+
+    logger = logging.getLogger("theano.gof.compilelock")
+    logger.setLevel(logging.ERROR)
+    logger = logging.getLogger("theano.tensor.opt")
+    logger.setLevel(logging.ERROR)
+    logger = logging.getLogger("exoplanet")
+    logger.setLevel(logging.DEBUG)
+
+    plt.style.use("default")
+    plt.rcParams["savefig.dpi"] = 100
+    plt.rcParams["figure.dpi"] = 100
+    plt.rcParams["font.size"] = 16
+    plt.rcParams["font.family"] = "sans-serif"
+    plt.rcParams["font.sans-serif"] = ["Liberation Sans"]
+    plt.rcParams["font.cursive"] = ["Liberation Sans"]
+    plt.rcParams["mathtext.fontset"] = "custom"
