+++ docs/tutorials/ipython_kernel_config.py
@@ -1,0 +1,2 @@
+c.InlineBackend.rc = {}
+c.InlineBackend.figure_format = "retina"

[…]
