### 2021-06-08 exoplanet-dev/exoplanet#196 @dfm
Done!
