### 2021-06-08 6ed21be
include jupytext dep
