+++ binder/requirements.txt
@@ -1,1 +1,2 @@
+jupytext
 .[docs]
