+++ binder/requirements.txt
@@ -1,2 +1,3 @@
+jupyterlab>=3.0
 jupytext
 .[docs]
