+++ binder/requirements.txt
@@ -1,3 +1,8 @@
 jupyterlab>=3.0
 jupytext
-.[docs]
+matplotlib
+corner
+lightkurve
+pymc3-ext>=0.1.0
+emcee
+.
