### 2021-06-08 b36794f
launch into classic mode for jupytext support
