### 2021-06-08 086fc88
Fixing "launch in binder" support for tutorials (#202)
* adding binder requirements

* include jupytext dep

* jupyterlab version

* explicit deps

* adding binder runtime

* launch into classic mode for jupytext support
