### 2021-06-10 f7ad980
adding note about multiprocessing (#203)
