### 2021-06-10 ddc35b3
Updates for JOSS review 1 (#186)
* make batman an optional dependency for tests

* fixing citation formatting

* add a few more tool references

* adding phoebe reference

* adding a figure
