+++ joss/paper.md
@@ -147,8 +147,8 @@ related.
   that must be compiled for performance. These include an efficient solver for
   Kepler's equation [based on the algorithm proposed by @raposo17] and limb
   darkened transit light curves [@agol20]. Besides the implementation for
-  `PyMC3` and `Theano`, `exoplanet-core` includes implementations in `numpy`
-  [@numpy] and `jax` [@jax].
+  `PyMC3`, `exoplanet-core` includes implementations in `numpy` [@numpy] and
+  `jax` [@jax].
 - `celerite2`[^celerite2], is an updated implementation of the _celerite_
   algorithm[^celerite] [@foremanmackey17; @foremanmackey18] for scalable
   Gaussian Process regression for time series data. Like `exoplanet-core`,
@@ -157,8 +157,8 @@ related.
 - `pymc3-ext`[^pymc3-ext], includes a set of helper functions to make `PyMC3`
   more amenable to the typical astronomical data analysis workflow. For example,
   it provides a tuning schedule for `PyMC3`'s sampler [based on the method used
-  by the `Stan` project @carpenter17] that provides better performance on models
-  with correlated parameters.
+  by the `Stan` project and described by @carpenter17] that provides better
+  performance on models with correlated parameters.
 - `rebound-pymc3`[^rebound-pymc3] provides an interface between _REBOUND_
   [@rein12], _REBOUNDx_ [@tamayo20], and `PyMC3` to enable inference with full
   N-body orbit integration.
@@ -183,6 +183,14 @@ automatically executed using GitHub Actions, but at lower cadence (once a week
 and when a new release of the `exoplanet` library is made) since the runtime is
 much longer.
 
+![Some examples of datasets fit using `exoplanet`. The full analyses behind
+these examples are available on the "Case Studies" page as Jupyter notebooks.
+(left) A fit to the light curves of a transiting exoplanet observed by two
+different space-based photometric surveys: Kepler and TESS. (right) The phase
+folded radial velocity time series for an exoplanet observed from different
+observatories with different instruments, fit simultaneously using `exoplanet`.
+\label{fig:figure}](figures/figure.png)
+
 # Similar tools
 
 There is a rich ecosystem of tooling available for inference with models such as
@@ -194,32 +202,35 @@ context.
 Some of the most popular tools in this space include (and note that this is far
 from a comprehensive list!) `EXOFAST` [@eastman13; @eastman19], `radvel`
 [@fulton18], `juliet` [@espinoza19], `exostriker` [@trifonov19], `PYANETI`
-[@barragan19], `allesfitter` [@guenther20], and `orbitize` [@blunt20]. These
-packages all focus on providing a high-level interface for designing models and
-then executing a fit. `exoplanet`, however, is designed to be lower level and
-more conceptually similar to tools like `batman` [@kreidberg15], `PyTransit`
-[@parviainen15], `ellc` [@maxted16], `starry` [@luger19], or `Limbdark.jl`
-[@agol20], which provide the building blocks for evaluating the models required
-for inference with exoplanet datasets. In fact, several of the higher-level
-packages listed above include these lower-level libraries as dependencies, and
-our hope is that `exoplanet` could provide the backend for future high-level
-libraries.
+[@barragan19], `allesfitter` [@guenther20], and `orbitize` [@blunt20]. Similar
+tools also exist for modeling observations of eclipsing binary systems,
+including `JKTEBOP` [@southworth04], `eb` [@irwin11], and `PHOEBE` [@conroy20].
+These packages all focus on providing a high-level interface for designing
+models and then executing a fit. `exoplanet`, however, is designed to be lower
+level and more conceptually similar to tools like `batman` [@kreidberg15],
+`PyTransit` [@parviainen15], `ldtk` [@parviainen15b], `ellc` [@maxted16],
[…]
