### 2021-06-10 a66c583
adding some inline comments and updating citations
