### 2021-06-10 e99d1bd
adding some inline comments and updating citations (#204)
