### 2021-06-10 #3285 @dfm
@benjaminpope, @grburgess: Thanks again so much for your thorough reviews! I've gone through and addressed all of your comments, I think. Here's [the project showing all the related issues and PRs](https://github.com/orgs/exoplanet-dev/projects/2). The big ones were:

- https://github.com/exoplanet-dev/exoplanet/pull/186: Outlines @benjaminpope's suggestions and links to specific issues or PRs for each one.
- https://github.com/exoplanet-dev/exoplanet/pull/200: Includes all the major updates to the tutorials on the main docs page.
- https://github.com/exoplanet-dev/case-studies/pull/21: Includes all of the infrastructure updates for the new case studies.

Overall the major changes were to the documentation, so:

- I completely overhauled the [documentation page](https://docs.exoplanet.codes/en/latest/) and [case studies](https://gallery.exoplanet.codes/). The big conceptual difference is that I moved some of the more specific tutorials to the case studies page and then added some more high level tutorials about [autodiff](https://docs.exoplanet.codes/en/latest/tutorials/autodiff/), [the range of models supported by the code](https://docs.exoplanet.codes/en/latest/tutorials/data-and-models/), and [some (limited) notes about reparameterization](https://docs.exoplanet.codes/en/latest/tutorials/reparameterization/).

- I updated the infrastructure to use the `sphinx-book-theme` for both the docs and case studies which includes buttons for launching the ipynbs using binder or downloading them directly. This should make it easier for users to dive right in without having to copy and paste the code and worry about missing imports.

- I dealt with (and tried to solve) various compatibility (ugh macos multiprocessing!) and runtime issues, added some more inline comments, and fixed various small bugs or issues that were identified.

Let me know if you see any remaining issues!!
