### 2021-06-10 #3285 @benjaminpope
@dfm - thanks for the updates. I have looked through the paper - happy with the new density of citations and the new figures, it's great. I have also looked at the new case studies and autodiff page - these are some of the most professional tutorial pages I have ever seen. I will be linking people to the autodiff one especially.

With that, all requested changes are complete and to a very high standard, and I am happy to recommend the paper for acceptance by JOSS.
