### 2021-06-11 #3285 @grburgess
@dfm I second acceptance. The docs look fantastic and very useful. Thanks for taking the time to make this project pedagogical as well as a beneficial to the community.
