### 2021-06-11 7064fb0
fixing a typo in the docs and adding reference to joss figure
