+++ docs/tutorials/citation.md
@@ -4,7 +4,7 @@ jupytext:
     extension: .md
     format_name: myst
     format_version: 0.13
-    jupytext_version: 1.11.2
+    jupytext_version: 1.11.3
 kernelspec:
   display_name: Python 3
   language: python
