### 2021-06-20 f42db86
[pre-commit.ci] pre-commit autoupdate (#206)
* [pre-commit.ci] pre-commit autoupdate

updates:
- [github.com/psf/black: 21.5b2 → 21.6b0](https://github.com/psf/black/compare/21.5b2...21.6b0)
- [github.com/mwouts/jupytext: v1.11.2 → v1.11.3](https://github.com/mwouts/jupytext/compare/v1.11.2...v1.11.3)

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
