### 2021-06-21 2ed4596
updating citation and changelog
