+++ .github/workflows/tests.yml
@@ -112,7 +112,7 @@ jobs:
   upload_pypi:
     needs: [tests, lint, build]
     runs-on: ubuntu-latest
-    if: github.event_name == 'release' && github.event.action == 'published'
+    if: startsWith(github.ref, 'refs/tags/')
     steps:
       - uses: actions/download-artifact@v2
         with:
