### 2021-06-21 45c662a
fixing direct dependencies for release
