### 2021-06-21 #3285 @dfm
@arfon: I've bumped the version number to `v0.5.1` and the archive is on Zenodo with the correct metadata at `10.5281/zenodo.5006965`. Thanks!
