+++ .pre-commit-config.yaml
@@ -14,7 +14,7 @@ repos:
     exclude: docs/tutorials
 
 - repo: https://github.com/PyCQA/isort
-  rev: "5.8.0"
+  rev: "5.9.1"
   hooks:
   - id: isort
     args: []
