### 2021-06-21 d194ee8
Adding test and fixing deprecated unit parameters
Co-authored-by: Daniel Yahalomi <danielyahalomi@gmail.com>
