### 2021-06-22 3dd400d
Adding test and fixing deprecated unit parameters (#211)
Co-authored-by: Daniel Yahalomi <danielyahalomi@gmail.com>

Co-authored-by: Daniel Yahalomi <danielyahalomi@gmail.com>
